start renaming parser modules to be more consistent
