+++ src/parse/converters/expressions/shared/refinement.js
@@ -1,7 +1,7 @@
 import { REFINEMENT } from 'config/types';
 import { expectedExpression } from './errors';
 import { name as namePattern } from './patterns';
-import getExpression from 'parse/converters/expression';
+import readExpression from 'parse/converters/readExpression';
 
 export default function getRefinement ( parser ) {
 	var start, name, expr;
@@ -28,7 +28,7 @@ export default function getRefinement ( parser ) {
 	if ( parser.matchString( '[' ) ) {
 		parser.allowWhitespace();
 
-		expr = getExpression( parser );
+		expr = readExpression( parser );
 		if ( !expr ) {
 			parser.error( expectedExpression );
 		}
