+++ src/parse/converters/readPartial.js
