+++ src/parse/converters/readText.js
@@ -1,7 +1,7 @@
 import getLowestIndex from './utils/getLowestIndex';
 import { decodeCharacterReferences } from 'utils/html';
 
-export default function ( parser ) {
+export default function readText ( parser ) {
 	var index, remaining, disallowed, barrier;
 
 	remaining = parser.remaining();
